Rename PredictionsView component to ClassificationsView
