+++ app/src/main/res/layout/component_classifications_view.xml
